Use "port_range_start/end" instead of udp_port for ICE range
mediasoup is not compatible with the UDP port multiplexing that LiveKit
uses when the `udp_port` setting is enabled, so it won't work anyways
and it's better to bt consistent between both Pion and mediasoup
engines.

Also extend the port range to 100 ports, given that now the multiplexing
feature won't be in use. 100 seems like a big enough range for local
development.